refactor: extracted user helper functions into test utils
